Cleaned up lint
